greg_format function
now in utils module.
