+++ mass_circular_weighing/routine_classes/results_summary_LaTeX.py
@@ -9,16 +9,8 @@ from tabulate import tabulate
 from msl.io import read
 
 from ..constants import IN_DEGREES_C, MU_STR
+from ..utils import greg_format
 from ..log import log
-
-
-def greg_format(number):
-    try:
-        before, after = '{:.9f}'.format(number).split('.')
-        return before + '.' + ' '.join(after[i:i+3] for i in range(0, len(after), 3))
-    except ValueError as e:
-        log.error("Couldn't put {} in Greg format. {}".format(number, e))
-        return str(number)
 
 
 def list_to_csstr(idlst):
